Swap lodash for lodash.template
Since this library is only using the template function of lodash, and
lodash publishes functions as individual npm packages, I swapped the
lodash for lodash.template. This reduced the bundled JS size from
37KB to 31KB, while maintaining all existing functionality. This also
has the added benefit of not bringing along the complete lodash library
into projects that are using a version of lodash with tree-shaking
configured.